de0 nano: swap GPIO core used, add sw tests, add input-only lines
Swap out the previously-used GPIO core for the OpenCores-based core.

Change config so the 8 LEDs are connected to the I/O lines, and 5 more pins (pushbutton, 4 DIP-switches) are accessible via this core.

Update the pin constraint files for this new layout.

Update top-level testbench.

Add some software for simulation and the board to test both basic GPIO features and interrupts.

New GPIO core is on the 32-bit bus arbiter, so the system's bus arbiters required a change.