or1k-lsu test improvements
This adds the "do the test once with tick timer exceptions
hammering" that many of the other tests do, since it has proven
to shake out extra bugs in many cases.

It also adds a test for "bursts" of writes smaller than
the word size, but into the same word.
e.g., assuming r1 is word aligned:
	l.sh	0(r1), r2
	l.sh	2(r1), r3