update dbus error tests for implementations with store buffer
When an implementation use a store buffer for writes, it will
continue execution after the store have been dispatched into the
store buffer until it finds a dependency on any of the writes
currently in it.
The dbus error exception can not be detected until the store
has propagated through the store buffer (i.e. when the actual
dbus access is performed).

The tests that force dbus errors with a store assumes that the
pipeline will stall until the dbus access for that particular
store have been performed.
To change that assumption, this patch adds a wait loop after
each dbus error forcing store. The wait loop checks for a flag
set in the exception vector.